+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Release: dd.mm.yyyy
 
 ### Bug Fixes
 
+- Fix an issue that prevented Phoenix from terminating after accessibility prompt ([#132](https://github.com/kasper/phoenix/issues/132)).
 - Fix an issue that prevented task callbacks from performing actions on the main thread ([#137](https://github.com/kasper/phoenix/issues/137)).
 
 ### API