+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Release: dd.mm.yyyy
 
 ### Changes
 
-- Breaking: Underscore.js is replaced with [Lodash](https://lodash.com) (4.16.6) ([#89](https://github.com/kasper/phoenix/issues/89)).
+- Breaking: Underscore.js is replaced with [Lodash](https://lodash.com) (4.17.3) ([#89](https://github.com/kasper/phoenix/issues/89)).
 - Xcode 8 is now required for building.
 
 ### Bug Fixes